Restructure responses
- Move shared responses module to common/responses
 - streaming.rs
 - get_response_impl and cancel_response_impl -> handlers.rs
- Move responses/ to regular/responses
