+++ sgl-router/src/routers/grpc/regular/responses/mod.rs
@@ -1,9 +1,8 @@
-//! gRPC Router `/v1/responses` endpoint implementation
+//! Regular gRPC Router `/v1/responses` endpoint implementation
 //!
-//! This module handles all responses-specific logic including:
+//! This module handles all responses-specific logic for the regular (non-Harmony) pipeline including:
 //! - Request validation
 //! - Conversation history and response chain loading
-//! - Background mode execution
 //! - Streaming support
 //! - MCP tool loop wrapper
 //! - Response persistence
@@ -12,11 +11,10 @@
 pub mod context;
 mod conversions;
 mod handlers;
-pub mod streaming;
 pub mod tool_loop;
 pub mod types;
 
 // Public exports
 pub use context::ResponsesContext;
-pub use handlers::{cancel_response_impl, get_response_impl, route_responses};
+pub use handlers::route_responses;
 pub use types::BackgroundTaskInfo;
